Refactor ShellContentBrowserClient into prod and test classes
