+++ cobalt/shell/browser/shell_content_browser_client.cc
@@ -121,13 +121,6 @@
 #include "services/network/public/mojom/ct_log_info.mojom.h"
 #endif
 
-#if defined(RUN_BROWSER_TESTS)
-#include "cobalt/shell/common/shell_controller.test-mojom.h"  // nogncheck
-#include "cobalt/shell/common/shell_test_switches.h"          // nogncheck
-#include "components/custom_handlers/simple_protocol_handler_registry_factory.h"  //nogncheck
-#include "components/metrics/test/test_enabled_state_provider.h"  // nogncheck
-#endif  // defined(RUN_BROWSER_TESTS)
-
 namespace content {
 
 namespace {
@@ -150,36 +143,8 @@ int GetCrashSignalFD(const base::CommandLine& command_line) {
   pid_t pid;
   return crash_reporter::GetHandlerSocket(&fd, &pid) ? fd : -1;
 }
+
 #endif
-
-#if defined(RUN_BROWSER_TESTS)
-class ShellControllerImpl : public mojom::ShellController {
- public:
-  ShellControllerImpl() = default;
-  ~ShellControllerImpl() override = default;
-
-  // mojom::ShellController:
-  void GetSwitchValue(const std::string& name,
-                      GetSwitchValueCallback callback) override {
-    const auto& command_line = *base::CommandLine::ForCurrentProcess();
-    if (command_line.HasSwitch(name)) {
-      std::move(callback).Run(command_line.GetSwitchValueASCII(name));
-    } else {
-      std::move(callback).Run(absl::nullopt);
-    }
-  }
-
-  void ExecuteJavaScript(const std::u16string& script,
-                         ExecuteJavaScriptCallback callback) override {
-    CHECK(!Shell::windows().empty());
-    WebContents* contents = Shell::windows()[0]->web_contents();
-    contents->GetPrimaryMainFrame()->ExecuteJavaScriptForTests(
-        script, std::move(callback));
-  }
-
-  void ShutDown() override { Shell::Shutdown(); }
-};
-#endif  // defined(RUN_BROWSER_TESTS)
 
 // TODO(crbug/1219642): Consider not needing VariationsServiceClient just to use
 // VariationsFieldTrialCreator.
@@ -240,10 +205,8 @@ base::flat_set<url::Origin> GetIsolatedContextOriginSetFromFlag() {
   std::string cmdline_origins(
       base::CommandLine::ForCurrentProcess()->GetSwitchValueASCII(
           switches::kIsolatedContextOrigins));
-
   std::vector<base::StringPiece> origin_strings = base::SplitStringPiece(
       cmdline_origins, ",", base::TRIM_WHITESPACE, base::SPLIT_WANT_NONEMPTY);
-
   base::flat_set<url::Origin> origin_set;
   for (const base::StringPiece& origin_string : origin_strings) {
     url::Origin allowed_origin = url::Origin::Create(GURL(origin_string));
@@ -256,18 +219,9 @@ base::flat_set<url::Origin> GetIsolatedContextOriginSetFromFlag() {
   return origin_set;
 }
 
-// In content browser tests we allow more than one ShellContentBrowserClient
-// to be created (actually, ContentBrowserTestContentBrowserClient). Any state
-// needed should be added here so that it's shared between the instances.
-struct SharedState {
-  SharedState() {}
+}  // namespace
 
-  // Owned by content::BrowserMainLoop.
-  raw_ptr<ShellBrowserMainParts, DanglingUntriaged> shell_browser_main_parts =
-      nullptr;
-
-  std::unique_ptr<PrefService> local_state;
-};
+SharedState::SharedState() {}
 
 SharedState& GetSharedState() {
   static SharedState* g_shared_state = nullptr;
@@ -276,27 +230,6 @@ SharedState& GetSharedState() {
   }
   return *g_shared_state;
 }
-
-#if defined(RUN_BROWSER_TESTS)
-std::unique_ptr<PrefService> CreateLocalState() {
-  auto pref_registry = base::MakeRefCounted<PrefRegistrySimple>();
-
-  metrics::MetricsService::RegisterPrefs(pref_registry.get());
-  variations::VariationsService::RegisterPrefs(pref_registry.get());
-
-  base::FilePath path;
-  CHECK(base::PathService::Get(SHELL_DIR_USER_DATA, &path));
-  path = path.AppendASCII("Local State");
-
-  PrefServiceFactory pref_service_factory;
-  pref_service_factory.set_user_prefs(
-      base::MakeRefCounted<JsonPrefStore>(path));
-
-  return pref_service_factory.Create(pref_registry);
-}
-#endif  // defined(RUN_BROWSER_TESTS)
-
-}  // namespace
 
 std::string GetShellUserAgent() {
   if (base::FeatureList::IsEnabled(blink::features::kFullUserAgent)) {
@@ -361,13 +294,6 @@ ShellContentBrowserClient::CreateBrowserMainParts(
 bool ShellContentBrowserClient::HasCustomSchemeHandler(
     content::BrowserContext* browser_context,
     const std::string& scheme) {
-#if defined(RUN_BROWSER_TESTS)
-  if (custom_handlers::ProtocolHandlerRegistry* protocol_handler_registry =
-          custom_handlers::SimpleProtocolHandlerRegistryFactory::
-              GetForBrowserContext(browser_context)) {
-    return protocol_handler_registry->IsHandledProtocol(scheme);
-  }
-#endif  // defined(RUN_BROWSER_TESTS)
   return false;
 }
 
@@ -379,17 +305,6 @@ ShellContentBrowserClient::CreateURLLoaderThrottles(
     NavigationUIData* navigation_ui_data,
     int frame_tree_node_id) {
   std::vector<std::unique_ptr<blink::URLLoaderThrottle>> result;
-
-#if defined(RUN_BROWSER_TESTS)
-  auto* factory = custom_handlers::SimpleProtocolHandlerRegistryFactory::
-      GetForBrowserContext(browser_context);
-  // null in unit tests.
-  if (factory) {
-    result.push_back(
-        std::make_unique<custom_handlers::ProtocolHandlerThrottle>(*factory));
-  }
-#endif  // defined(RUN_BROWSER_TESTS)
-
   return result;
 }
 
@@ -421,17 +336,6 @@ void ShellContentBrowserClient::AppendExtraCommandLineSwitches(
 
   command_line->CopySwitchesFrom(*base::CommandLine::ForCurrentProcess(),
                                  kForwardSwitches, std::size(kForwardSwitches));
-
-#if defined(RUN_BROWSER_TESTS)
-  static const char* kForwardTestSwitches[] = {
-      switches::kExposeInternalsForTesting,
-      switches::kRunWebTests,
-  };
-
-  command_line->CopySwitchesFrom(*base::CommandLine::ForCurrentProcess(),
-                                 kForwardTestSwitches,
-                                 std::size(kForwardTestSwitches));
-#endif  // defined(RUN_BROWSER_TESTS)
 
 #if BUILDFLAG(IS_LINUX)
   if (base::CommandLine::ForCurrentProcess()->HasSwitch(
@@ -701,18 +605,6 @@ BluetoothDelegate* ShellContentBrowserClient::GetBluetoothDelegate() {
 }
 #endif
 
-#if defined(RUN_BROWSER_TESTS)
-void ShellContentBrowserClient::BindBrowserControlInterface(
-    mojo::ScopedMessagePipeHandle pipe) {
-  if (!pipe.is_valid()) {
-    return;
-  }
-  mojo::MakeSelfOwnedReceiver(
-      std::make_unique<ShellControllerImpl>(),
-      mojo::PendingReceiver<mojom::ShellController>(std::move(pipe)));
-}
-#endif  // defined(RUN_BROWSER_TESTS)
-
 void ShellContentBrowserClient::set_browser_main_parts(
     ShellBrowserMainParts* parts) {
   GetSharedState().shell_browser_main_parts = parts;
@@ -766,90 +658,9 @@ bool ShellContentBrowserClient::HasErrorPage(int http_status_code) {
   return http_status_code >= 400 && http_status_code < 600;
 }
 
-void ShellContentBrowserClient::CreateFeatureListAndFieldTrials() {
-#if defined(RUN_BROWSER_TESTS)
-  GetSharedState().local_state = CreateLocalState();
-  SetUpFieldTrials();
-  // Schedule a Local State write since the above function resulted in some
-  // prefs being updated.
-  GetSharedState().local_state->CommitPendingWrite();
-#endif  // defined(RUN_BROWSER_TESTS)
-}
+void ShellContentBrowserClient::CreateFeatureListAndFieldTrials() {}
 
-void ShellContentBrowserClient::SetUpFieldTrials() {
-#if defined(RUN_BROWSER_TESTS)
-  metrics::TestEnabledStateProvider enabled_state_provider(/*consent=*/false,
-                                                           /*enabled=*/false);
-  base::FilePath path;
-  base::PathService::Get(SHELL_DIR_USER_DATA, &path);
-  std::unique_ptr<metrics::MetricsStateManager> metrics_state_manager =
-      metrics::MetricsStateManager::Create(
-          GetSharedState().local_state.get(), &enabled_state_provider,
-          std::wstring(), path, metrics::StartupVisibility::kUnknown,
-          {
-              .force_benchmarking_mode =
-                  base::CommandLine::ForCurrentProcess()->HasSwitch(
-                      cc::switches::kEnableGpuBenchmarking),
-          });
-  metrics_state_manager->InstantiateFieldTrialList();
-
-  std::vector<std::string> variation_ids;
-  auto feature_list = std::make_unique<base::FeatureList>();
-
-  std::unique_ptr<variations::SeedResponse> initial_seed;
-#if BUILDFLAG(IS_ANDROID)
-  if (!GetSharedState().local_state->HasPrefPath(
-          variations::prefs::kVariationsSeedSignature)) {
-    DVLOG(1) << "Importing first run seed from Java preferences.";
-    initial_seed = variations::android::GetVariationsFirstRunSeed();
-  }
-#endif
-
-  ShellVariationsServiceClient variations_service_client;
-  variations::VariationsFieldTrialCreator field_trial_creator(
-      &variations_service_client,
-      std::make_unique<variations::VariationsSeedStore>(
-          GetSharedState().local_state.get(), std::move(initial_seed),
-          /*signature_verification_enabled=*/true),
-      variations::UIStringOverrider());
-
-  variations::SafeSeedManager safe_seed_manager(
-      GetSharedState().local_state.get());
-
-  const base::CommandLine& command_line =
-      *base::CommandLine::ForCurrentProcess();
-
-  // Overrides for content/common and lower layers' switches.
-  std::vector<base::FeatureList::FeatureOverrideInfo> feature_overrides =
-      content::GetSwitchDependentFeatureOverrides(command_line);
-
-  // Overrides for content/shell switches.
-
-  // Overrides for --run-web-tests.
-  if (switches::IsRunWebTestsSwitchPresent()) {
-    // Disable artificial timeouts for PNA-only preflights in warning-only mode
-    // for web tests. We do not exercise this behavior with web tests as it is
-    // intended to be a temporary rollout stage, and the short timeout causes
-    // flakiness when the test server takes just a tad too long to respond.
-    feature_overrides.emplace_back(
-        std::cref(
-            network::features::kPrivateNetworkAccessPreflightShortTimeout),
-        base::FeatureList::OVERRIDE_DISABLE_FEATURE);
-  }
-
-  // Since this is a test-only code path, some arguments to SetUpFieldTrials are
-  // null.
-  // TODO(crbug/1248066): Consider passing a low entropy source.
-  variations::PlatformFieldTrials platform_field_trials;
-  field_trial_creator.SetUpFieldTrials(
-      variation_ids,
-      command_line.GetSwitchValueASCII(
-          variations::switches::kForceVariationIds),
-      feature_overrides, std::move(feature_list), metrics_state_manager.get(),
-      &platform_field_trials, &safe_seed_manager,
-      /*add_entropy_source_to_variations_ids=*/false);
-#endif  // defined(RUN_BROWSER_TESTS)
-}
+void ShellContentBrowserClient::SetUpFieldTrials() {}
 
 absl::optional<blink::ParsedPermissionsPolicy>
 ShellContentBrowserClient::GetPermissionsPolicyForIsolatedWebApp(
