+++ cobalt/testing/browser_tests/browser/shell_content_browser_test_client.cc
@@ -1,0 +1,281 @@
+// Copyright 2025 The Cobalt Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "cobalt/testing/browser_tests/browser/shell_content_browser_test_client.h"
+
+#include <string>
+#include <utility>
+#include <vector>
+
+#include "base/base_switches.h"
+#include "base/command_line.h"
+#include "base/logging.h"
+#include "base/path_service.h"
+#include "base/strings/string_number_conversions.h"
+#include "build/build_config.h"
+#include "cc/base/switches.h"
+#include "cobalt/shell/browser/shell.h"
+#include "cobalt/shell/browser/shell_paths.h"
+#include "cobalt/shell/common/shell_controller.test-mojom.h"
+#include "cobalt/shell/common/shell_switches.h"
+#include "cobalt/testing/browser_tests/common/shell_test_switches.h"
+#include "components/custom_handlers/protocol_handler_registry.h"
+#include "components/custom_handlers/protocol_handler_throttle.h"
+#include "components/custom_handlers/simple_protocol_handler_registry_factory.h"
+#include "components/metrics/metrics_service.h"
+#include "components/metrics/metrics_state_manager.h"
+#include "components/metrics/test/test_enabled_state_provider.h"
+#include "components/prefs/json_pref_store.h"
+#include "components/prefs/pref_registry_simple.h"
+#include "components/prefs/pref_service_factory.h"
+#include "components/variations/platform_field_trials.h"
+#include "components/variations/pref_names.h"
+#include "components/variations/service/safe_seed_manager.h"
+#include "components/variations/service/variations_field_trial_creator.h"
+#include "components/variations/service/variations_service.h"
+#include "components/variations/service/variations_service_client.h"
+#include "components/variations/variations_switches.h"
+#include "content/public/browser/web_contents.h"
+#include "content/public/common/content_switch_dependent_feature_overrides.h"
+#include "mojo/public/cpp/bindings/self_owned_receiver.h"
+#include "services/network/public/cpp/features.h"
+
+#if BUILDFLAG(IS_ANDROID)
+#include "components/variations/android/variations_seed_bridge.h"
+#endif
+
+#if BUILDFLAG(IS_LINUX) || BUILDFLAG(IS_ANDROID)
+#include "components/crash/core/app/crash_switches.h"
+#include "components/crash/core/app/crashpad.h"
+#include "content/public/common/content_descriptors.h"
+#endif
+
+namespace content {
+
+namespace {
+
+// TODO(crbug/1219642): Consider not needing VariationsServiceClient just to use
+// VariationsFieldTrialCreator.
+class ShellVariationsServiceClient
+    : public variations::VariationsServiceClient {
+ public:
+  ShellVariationsServiceClient() = default;
+  ~ShellVariationsServiceClient() override = default;
+
+  // variations::VariationsServiceClient:
+  base::Version GetVersionForSimulation() override { return base::Version(); }
+  scoped_refptr<network::SharedURLLoaderFactory> GetURLLoaderFactory()
+      override {
+    return nullptr;
+  }
+  network_time::NetworkTimeTracker* GetNetworkTimeTracker() override {
+    return nullptr;
+  }
+  version_info::Channel GetChannel() override {
+    return version_info::Channel::UNKNOWN;
+  }
+  bool OverridesRestrictParameter(std::string* parameter) override {
+    return false;
+  }
+  bool IsEnterprise() override { return false; }
+  // Profiles aren't supported, so nothing to do here.
+  void RemoveGoogleGroupsFromPrefsForDeletedProfiles(
+      PrefService* local_state) override {}
+};
+
+class ShellControllerImpl : public mojom::ShellController {
+ public:
+  ShellControllerImpl() = default;
+  ~ShellControllerImpl() override = default;
+
+  // mojom::ShellController:
+  void GetSwitchValue(const std::string& name,
+                      GetSwitchValueCallback callback) override {
+    const auto& command_line = *base::CommandLine::ForCurrentProcess();
+    if (command_line.HasSwitch(name)) {
+      std::move(callback).Run(command_line.GetSwitchValueASCII(name));
+    } else {
+      std::move(callback).Run(absl::nullopt);
+    }
+  }
+
+  void ExecuteJavaScript(const std::u16string& script,
+                         ExecuteJavaScriptCallback callback) override {
+    CHECK(!Shell::windows().empty());
+    WebContents* contents = Shell::windows()[0]->web_contents();
+    contents->GetPrimaryMainFrame()->ExecuteJavaScriptForTests(
+        script, std::move(callback));
+  }
+
+  void ShutDown() override { Shell::Shutdown(); }
+};
+
+std::unique_ptr<PrefService> CreateLocalState() {
+  auto pref_registry = base::MakeRefCounted<PrefRegistrySimple>();
+
+  metrics::MetricsService::RegisterPrefs(pref_registry.get());
+  variations::VariationsService::RegisterPrefs(pref_registry.get());
+
+  base::FilePath path;
+  CHECK(base::PathService::Get(SHELL_DIR_USER_DATA, &path));
+  path = path.AppendASCII("Local State");
+
+  PrefServiceFactory pref_service_factory;
+  pref_service_factory.set_user_prefs(
+      base::MakeRefCounted<JsonPrefStore>(path));
+
+  return pref_service_factory.Create(pref_registry);
+}
+
+}  // namespace
+
+ShellContentBrowserTestClient::ShellContentBrowserTestClient() = default;
+ShellContentBrowserTestClient::~ShellContentBrowserTestClient() = default;
+
+void ShellContentBrowserTestClient::BindBrowserControlInterface(
+    mojo::ScopedMessagePipeHandle pipe) {
+  if (!pipe.is_valid()) {
+    return;
+  }
+  mojo::MakeSelfOwnedReceiver(
+      std::make_unique<ShellControllerImpl>(),
+      mojo::PendingReceiver<mojom::ShellController>(std::move(pipe)));
+}
+
+bool ShellContentBrowserTestClient::HasCustomSchemeHandler(
+    content::BrowserContext* browser_context,
+    const std::string& scheme) {
+  if (custom_handlers::ProtocolHandlerRegistry* protocol_handler_registry =
+          custom_handlers::SimpleProtocolHandlerRegistryFactory::
+              GetForBrowserContext(browser_context)) {
+    return protocol_handler_registry->IsHandledProtocol(scheme);
+  }
+  return false;
+}
+
+std::vector<std::unique_ptr<blink::URLLoaderThrottle>>
+ShellContentBrowserTestClient::CreateURLLoaderThrottles(
+    const network::ResourceRequest& request,
+    BrowserContext* browser_context,
+    const base::RepeatingCallback<WebContents*()>& wc_getter,
+    NavigationUIData* navigation_ui_data,
+    int frame_tree_node_id) {
+  std::vector<std::unique_ptr<blink::URLLoaderThrottle>> result;
+
+  auto* factory = custom_handlers::SimpleProtocolHandlerRegistryFactory::
+      GetForBrowserContext(browser_context);
+  if (factory) {
+    result.push_back(
+        std::make_unique<custom_handlers::ProtocolHandlerThrottle>(*factory));
+  }
+  return result;
+}
+
+void ShellContentBrowserTestClient::AppendExtraCommandLineSwitches(
+    base::CommandLine* command_line,
+    int child_process_id) {
+  ShellContentBrowserClient::AppendExtraCommandLineSwitches(command_line,
+                                                            child_process_id);
+  static const char* kForwardTestSwitches[] = {
+      test_switches::kExposeInternalsForTesting,
+      test_switches::kRunWebTests,
+  };
+
+  command_line->CopySwitchesFrom(*base::CommandLine::ForCurrentProcess(),
+                                 kForwardTestSwitches,
+                                 std::size(kForwardTestSwitches));
+}
+
+void ShellContentBrowserTestClient::CreateFeatureListAndFieldTrials() {
+  GetSharedState().local_state = CreateLocalState();
+  SetUpFieldTrials();
+  // Schedule a Local State write since the above function resulted in some
+  // prefs being updated.
+  GetSharedState().local_state->CommitPendingWrite();
+}
+
+void ShellContentBrowserTestClient::SetUpFieldTrials() {
+  metrics::TestEnabledStateProvider enabled_state_provider(/*consent=*/false,
+                                                           /*enabled=*/false);
+  base::FilePath path;
+  base::PathService::Get(SHELL_DIR_USER_DATA, &path);
+  std::unique_ptr<metrics::MetricsStateManager> metrics_state_manager =
+      metrics::MetricsStateManager::Create(
+          GetSharedState().local_state.get(), &enabled_state_provider,
+          std::wstring(), path, metrics::StartupVisibility::kUnknown,
+          {
+              .force_benchmarking_mode =
+                  base::CommandLine::ForCurrentProcess()->HasSwitch(
+                      cc::switches::kEnableGpuBenchmarking),
+          });
+  metrics_state_manager->InstantiateFieldTrialList();
+
+  std::vector<std::string> variation_ids;
+  auto feature_list = std::make_unique<base::FeatureList>();
+
+  std::unique_ptr<variations::SeedResponse> initial_seed;
+#if BUILDFLAG(IS_ANDROID)
+  if (!GetSharedState().local_state->HasPrefPath(
+          variations::prefs::kVariationsSeedSignature)) {
+    DVLOG(1) << "Importing first run seed from Java preferences.";
+    initial_seed = variations::android::GetVariationsFirstRunSeed();
+  }
+#endif
+
+  ShellVariationsServiceClient variations_service_client;
+  variations::VariationsFieldTrialCreator field_trial_creator(
+      &variations_service_client,
+      std::make_unique<variations::VariationsSeedStore>(
+          GetSharedState().local_state.get(), std::move(initial_seed),
+          /*signature_verification_enabled=*/true),
+      variations::UIStringOverrider());
+
+  variations::SafeSeedManager safe_seed_manager(
+      GetSharedState().local_state.get());
+
+  const base::CommandLine& command_line =
+      *base::CommandLine::ForCurrentProcess();
+
+  // Overrides for cobalt/common and lower layers' switches.
+  std::vector<base::FeatureList::FeatureOverrideInfo> feature_overrides =
+      content::GetSwitchDependentFeatureOverrides(command_line);
+
+  // Overrides for cobalt/shell switches.
+
+  // Overrides for --run-web-tests.
+  if (test_switches::IsRunWebTestsSwitchPresent()) {
+    // Disable artificial timeouts for PNA-only preflights in warning-only mode
+    // for web tests. We do not exercise this behavior with web tests as it is
+    // intended to be a temporary rollout stage, and the short timeout causes
+    // flakiness when the test server takes just a tad too long to respond.
+    feature_overrides.emplace_back(
+        std::cref(
+            network::features::kPrivateNetworkAccessPreflightShortTimeout),
+        base::FeatureList::OVERRIDE_DISABLE_FEATURE);
+  }
+
+  // Since this is a test-only code path, some arguments to SetUpFieldTrials are
+  // null.
+  // TODO(crbug/1248066): Consider passing a low entropy source.
+  variations::PlatformFieldTrials platform_field_trials;
+  field_trial_creator.SetUpFieldTrials(
+      variation_ids,
+      command_line.GetSwitchValueASCII(
+          variations::switches::kForceVariationIds),
+      feature_overrides, std::move(feature_list), metrics_state_manager.get(),
+      &platform_field_trials, &safe_seed_manager,
+      /*add_entropy_source_to_variations_ids=*/false);
+}
+
+}  // namespace content
